+++ lib/src/validators/string_validators.dart
@@ -33,14 +33,16 @@ Validator<String> hasMinUppercaseChars({
   }
 
   return (String input) {
-    int uppercaseCount = customUppercaseCounter?.call(input) ??
+    int uppercaseCount =
+        customUppercaseCounter?.call(input) ??
         _upperAndLowerCaseCounter(input).upperCount;
 
     return uppercaseCount >= min
         ? null
         : hasMinUppercaseCharsMsg?.call(input, min) ??
-            FormBuilderLocalizations.current
-                .containsUppercaseCharErrorText(min);
+              FormBuilderLocalizations.current.containsUppercaseCharErrorText(
+                min,
+              );
   };
 }
 
@@ -54,14 +56,16 @@ Validator<String> hasMinLowercaseChars({
     throw ArgumentError.value(min, 'min', 'This argument must be at least 1');
   }
   return (String input) {
-    int lowercaseCount = customLowercaseCounter?.call(input) ??
+    int lowercaseCount =
+        customLowercaseCounter?.call(input) ??
         _upperAndLowerCaseCounter(input).lowerCount;
 
     return lowercaseCount >= min
         ? null
         : hasMinLowercaseCharsMsg?.call(input, min) ??
-            FormBuilderLocalizations.current
-                .containsLowercaseCharErrorText(min);
+              FormBuilderLocalizations.current.containsLowercaseCharErrorText(
+                min,
+              );
   };
 }
 
@@ -75,12 +79,13 @@ Validator<String> hasMinNumericChars({
     throw ArgumentError.value(min, 'min', 'This argument must be at least 1');
   }
   return (String input) {
-    final int numericCount = customNumericCounter?.call(input) ??
+    final int numericCount =
+        customNumericCounter?.call(input) ??
         _numericRegex.allMatches(input).length;
     return numericCount >= min
         ? null
         : hasMinNumericCharsMsg?.call(input, min) ??
-            FormBuilderLocalizations.current.containsNumberErrorText(min);
+              FormBuilderLocalizations.current.containsNumberErrorText(min);
   };
 }
 
@@ -98,7 +103,8 @@ Validator<String> hasMinSpecialChars({
     if (customSpecialCounter == null) {
       final (lowerCount: int lowerCount, upperCount: int upperCount) =
           _upperAndLowerCaseCounter(input);
-      specialCount = input.length -
+      specialCount =
+          input.length -
           (lowerCount + upperCount + _numericRegex.allMatches(input).length);
     } else {
       specialCount = customSpecialCounter.call(input);
@@ -107,7 +113,9 @@ Validator<String> hasMinSpecialChars({
     return specialCount >= min
         ? null
         : hasMinSpecialCharsMsg?.call(input, min) ??
-            FormBuilderLocalizations.current.containsSpecialCharErrorText(min);
+              FormBuilderLocalizations.current.containsSpecialCharErrorText(
+                min,
+              );
   };
 }
 
@@ -120,7 +128,7 @@ Validator<String> match(
     return regExp.hasMatch(input)
         ? null
         : matchMsg?.call(input) ??
-            FormBuilderLocalizations.current.matchErrorText;
+              FormBuilderLocalizations.current.matchErrorText;
   };
 }
 
@@ -137,7 +145,7 @@ Validator<String> uuid({
     return (regex ?? _uuidRegex).hasMatch(input)
         ? null
         : uuidMsg?.call(input) ??
-            FormBuilderLocalizations.current.uuidErrorText;
+              FormBuilderLocalizations.current.uuidErrorText;
   };
 }
 
@@ -147,15 +155,56 @@ Validator<String> contains(
   bool caseSensitive = true,
   String Function(String substring, String input)? containsMsg,
 }) {
+  final String lowercaseSubstring = substring.toLowerCase();
   return (String input) {
     if (substring.isEmpty) {
       return null;
     } else if (caseSensitive
         ? input.contains(substring)
-        : input.toLowerCase().contains(substring.toLowerCase())) {
+        : input.toLowerCase().contains(lowercaseSubstring)) {
       return null;
     }
     return containsMsg?.call(substring, input) ??
         FormBuilderLocalizations.current.containsErrorText(substring);
+  };
+}
+
+/// {@macro validator_starts_with}
+Validator<String> startsWith(
+  String prefix, {
+  bool caseSensitive = true,
+  String Function(String prefix, String input)? startsWithMsg,
+}) {
+  final String lowercasePrefix = prefix.toLowerCase();
+  return (String input) {
+    if (prefix.isEmpty) {
+      return null;
+    } else if (caseSensitive
+        ? input.startsWith(prefix)
+        : input.toLowerCase().startsWith(lowercasePrefix)) {
+      return null;
+    }
+    return startsWithMsg?.call(prefix, input) ??
+        FormBuilderLocalizations.current.startsWithErrorText(prefix);
+  };
+}
+
+/// {@macro validator_ends_with}
+Validator<String> endsWith(
+  String suffix, {
+  bool caseSensitive = true,
+  String Function(String suffix, String input)? endsWithMsg,
+}) {
+  final String lowercaseSuffix = suffix.toLowerCase();
+  return (String input) {
+    if (suffix.isEmpty) {
+      return null;
+    } else if (caseSensitive
+        ? input.endsWith(suffix)
+        : input.toLowerCase().endsWith(lowercaseSuffix)) {
+      return null;
+    }
+    return endsWithMsg?.call(suffix, input) ??
+        FormBuilderLocalizations.current.endsWithErrorText(suffix);
   };
 }